requirements.txt
celery==5.2.3
Django==3.2
django-cors-headers==3.11.0
django-filter==21.1
django-resized==0.3.11
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
fcm-django==1.0.8
psycopg2-binary==2.9.3
python-dotenv==0.19.2
redis==4.1.2
zeep==4.1.0